% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/habgam.R
\docType{data}
\name{habgam}
\alias{habgam}
\title{Generalized additive model for habitat}
\format{A \code{\link[mgcv]{gam}} model}
\usage{
habgam
}
\description{
Geenralized additive model for habitat
}
\examples{
data(habgam)
}
\keyword{datasets}
