% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/habgamwp.R
\docType{data}
\name{habgamwp}
\alias{habgamwp}
\title{Generalized additive model for habitat}
\format{A \code{\link[mgcv]{gam}} model}
\usage{
habgamwp
}
\description{
Geenralized additive model for habitat
}
\examples{
data(habgamwp)
}
\keyword{datasets}
