% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/habglm.R
\docType{data}
\name{habglm}
\alias{habglm}
\title{Generalized linear model for habitat}
\format{A \code{\link[stats]{glm}} model}
\usage{
habglm
}
\description{
Geenralized linear model for habitat
}
\examples{
data(habglm)
}
\keyword{datasets}
