% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/habrf.R
\docType{data}
\name{habrf}
\alias{habrf}
\title{Random forest model for habitat}
\format{A \code{\link[randomForest]{randomForest}} model}
\usage{
habrf
}
\description{
Random forest model for habitat
}
\examples{
data(habrf)
}
\keyword{datasets}
