% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/habrfwp.R
\docType{data}
\name{habrfwp}
\alias{habrfwp}
\title{Random forest model for habitat}
\format{A \code{\link[randomForest]{randomForest}} model}
\usage{
habrfwp
}
\description{
Random forest model for habitat
}
\examples{
data(habrfwp)
}
\keyword{datasets}
