% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wqgam.R
\docType{data}
\name{wqgam}
\alias{wqgam}
\title{Generalized additive model for water quality}
\format{A \code{\link[mgcv]{gam}} model}
\usage{
wqgam
}
\description{
Geenralized additive model for water quality
}
\examples{
data(wqgam)
}
\keyword{datasets}
