% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wqgamwp.R
\docType{data}
\name{wqgamwp}
\alias{wqgamwp}
\title{Generalized additive model for water quality}
\format{A \code{\link[mgcv]{gam}} model}
\usage{
wqgamwp
}
\description{
Geenralized additive model for water quality
}
\examples{
data(wqgamwp)
}
\keyword{datasets}
