% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wqglm.R
\docType{data}
\name{wqglm}
\alias{wqglm}
\title{Generalized linear model for water quality}
\format{A \code{\link[stats]{glm}} model}
\usage{
wqglm
}
\description{
Geenralized linear model for water quality
}
\examples{
data(wqglm)
}
\keyword{datasets}
