% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wqi.R
\name{wqi}
\alias{wqi}
\title{WQI}
\usage{
wqi(datin)
}
\arguments{
\item{datin}{input \code{data.frame} with chemical and biological data}
}
\value{
a \code{data.frame} same as \code{datin} but with new columns for \code{pChem}, \code{pHab}, \code{pChemHab}, \code{BiologicalCondtion}, \code{WaterChemistryCondition}, \code{HabitatCondition}, \code{OverallStressCondition}, and \code{StreamHealthIndex}
}
\description{
Water quality index function
}
\details{
See \code{\link{sampdat}} for required input format
}
\examples{
wqi(sampdat)
}
