% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wqrf.R
\docType{data}
\name{wqrf}
\alias{wqrf}
\title{Random forest model for water quality}
\format{A \code{\link[randomForest]{randomForest}} model}
\usage{
wqrf
}
\description{
Random forest model for water quality
}
\examples{
data(wqrf)
}
\keyword{datasets}
