% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wqrfwp.R
\docType{data}
\name{wqrfwp}
\alias{wqrfwp}
\title{Random forest model for water quality}
\format{A \code{\link[randomForest]{randomForest}} model}
\usage{
wqrfwp
}
\description{
Random forest model for water quality
}
\examples{
data(wqrfwp)
}
\keyword{datasets}
