% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sqi.R
\name{sqi}
\alias{sqi}
\title{SQI}
\usage{
sqi(datin, wq_mod_in = NULL, hab_mod_in = NULL, lothrsh = 0.1,
  hithrsh = 0.9)
}
\arguments{
\item{datin}{input \code{data.frame} with chemical, physical, and biological data}

\item{wq_mod_in}{input model object for predicting stressed state for water quality}

\item{hab_mod_in}{input model object for prediction stressed state for habitat}

\item{lothrsh}{numeric indicating lower threshold for low stress}

\item{hithrsh}{numeric indicating upper threshold for high stress}
}
\value{
a \code{data.frame} same as \code{datin} but with new columns for \code{pChem}, \code{pHab}, \code{pChemHab}, \code{BiologicalCondtion}, \code{WaterChemistryCondition}, \code{HabitatCondition}, \code{OverallStressCondition}, and \code{StreamHealthIndex}
}
\description{
Stream quality index function
}
\details{
See \code{\link{sampdat}} for required input format.  \code{wq_mod_in} and \code{hab_mod_in} are both \code{\link[stats]{glm}} objects (\code{\link{wqglm}} and \code{\link{habglm}}) included with the package.
}
\examples{

sqi(sampdat)

}
