% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strs_surf.R
\name{strs_surf}
\alias{strs_surf}
\title{Plot probability surface for likelihood of stress}
\usage{
strs_surf(xvar, yvar, mod = c("hab_mod", "wq_mod"), title = TRUE,
  lenv = 200, opt_vrs = NULL, low = "#d7191c", mid = "#ffffbf",
  high = "#2c7bb6")
}
\arguments{
\item{xvar}{chr string of variable for x-axis}

\item{yvar}{chr string of variable for y-axix}

\item{mod}{chr string of habitat or water quality model to use, must be either \code{'hab_mod'} or \code{'wq_mod'}}

\item{title}{logical indicating if title is shown above plot}

\item{lenv}{numeric indicating resolution of probability surface}

\item{opt_vrs}{names list with optional constant values for variables not specified by \code{xvar} and \code{yvar}}

\item{low}{chr string for color on low end of palette}

\item{mid}{chr string for color at mid range of palette}

\item{high}{chr string for color at high end of palette}
}
\value{
A \code{\link[ggplot2]{ggplot}} object
}
\description{
Plot probability surface for likelihood of habitat or water quality stress
}
\details{
Plots a smooth surface of the likelihood of either habitat or water quality stress for selected ranges of two variables.  The remaining variables are held constant at average values for those in the calibration dataset unless manually set with user input.
}
\examples{

# water quality stress
opt_vrs <- list(
  TN2 = 2.22,
  TP = 0.156,
  Cond = 1139
)

strs_surf(xvar = 'TN2', yvar = 'TP', mod = 'wq_mod', opt_vrs = opt_vrs)

# habitat stress
opt_vrs <- list(
   indexscore_cram = 100,
   PCT_SAFN = 25,
   H_AqHab = 0,
   H_SubNat = 1.3,
   Ev_FlowHab = 0.17,
   XCMG = 100
 )
 
strs_surf(xvar = 'PCT_SAFN', yvar = 'indexscore_cram', mod = 'hab_mod', opt_vrs = opt_vrs)
}
