% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wqi.R
\name{wqi}
\alias{wqi}
\title{WQI}
\usage{
wqi(datin, wq_mod_in = NULL, hab_mod_in = NULL)
}
\arguments{
\item{datin}{input \code{data.frame} with chemical and biological data}

\item{wq_mod_in}{input model object for predicting stressed state for water quality}

\item{hab_mod_in}{input model object for prediction stressed state for habitat}
}
\value{
a \code{data.frame} same as \code{datin} but with new columns for \code{pChem}, \code{pHab}, \code{pChemHab}, \code{BiologicalCondtion}, \code{WaterChemistryCondition}, \code{HabitatCondition}, \code{OverallStressCondition}, and \code{StreamHealthIndex}
}
\description{
Water quality index function
}
\details{
See \code{\link{sampdat}} for required input format.  \code{wq_mod_in} and \code{hab_mod_in} are both \code{\link[randomForest]{randomForest}} objects (\code{\link{wqrf}} and \code{\link{habrf}}, default) or \code{\link[mgcv]{gam}} objects (\code{\link{wqgam}} and \code{\link{habgam}}) included with the package.
}
\examples{

# using random forest models (default)
wqi(sampdat)

# using GAMs
wqi(sampdat, wq_mod_in = wqgam, hab_mod_in = habgam)
}
